+++ robotcode/language_server/robotframework/parts/hover.py
@@ -20,7 +20,7 @@ from ....utils.logging import LoggingDescriptor
 from ...common.decorators import language_id
 from ...common.lsp_types import Hover, MarkupContent, MarkupKind, Position
 from ...common.text_document import TextDocument
-from ..utils.ast import (
+from ..utils.ast_utils import (
     HasTokens,
     Token,
     get_nodes_at_position,