+++ CHANGELOG.md
@@ -2,11 +2,14 @@
 
 All notable changes to the "robotcode" extension will be documented in this file.
 
-## [Unreleased]
-- none so far
+## [Unreleased]
+
+- Provide better error messages if python and robot environment not matches RobotCode requirements
+  - fixes [#40](https://github.com/d-biehl/robotcode/issues/40)
+
 ##  0.11.0
 
-- correct find references at token ends
+- Correct find references at token ends
   - If the cursor is at the end of a keyword, for example, the keyword will also be highlighted and the references will be found.
 
 ##  0.10.1